fix: fix cr comments

+++ MergerLogic/DataTypes/Fs.cs
@@ -12,7 +12,7 @@ namespace MergerLogic.DataTypes
 
         private IEnumerator<Tile> _tiles;
         private bool _done;
-        private int _completedTiles;
+        private long _completedTiles;
 
         private readonly IPathUtils _pathUtils;
         private readonly IFileSystem _fileSystem;
@@ -92,10 +92,10 @@ namespace MergerLogic.DataTypes
 
         public override void setBatchIdentifier(string batchIdentifier)
         {
-            this._completedTiles = int.Parse(batchIdentifier);
+            this._completedTiles = long.Parse(batchIdentifier);
             // uncomment to make this function work at any point of the run and not only after the source initialization
             //this.tiles.Reset();
-            for (int i = 0; i < this._completedTiles; i++)
+            for (long i = 0; i < this._completedTiles; i++)
             {
                 this._tiles.MoveNext();
             }
